Receiver: [FIX] improve stability
The receiver used to have issues when a report sender connected and then
did not close the connection. When too many senders showed this
behaviour no new connections where possible. To reduce the severity of
this issue I implemented a timeout for sending reports and increased the
number of servers. This still does not fix the underlying problem but
buys us a little more time.