Receiver: TAP archive support
It now recognizes if the TAP stream is in fact a mime type
of a compressed tar and then marks this accordingly in db.

The content is still stored in the same db field 'tap'
which originally contained the TAP. The fact that it is
maybe an archive needs to be evaluated always from now on.
This is already implemented in Artemis::TAP::Harness.