Reports-Rcv: port to AnyEvent
The Tapper reports receiver was based on Net::Server::Fork. As the name
implies this tool forks a new process for every incoming connection. To reduce
the performance impact of this approach Net::Server::Fork enforces an upper
limit on the number of simulataneously running child processes.

This did not meet our needs and thus I ported the reports receiver to use an
event based framework. This should reduce the overhead and dispose the upper
limit on concurrent requests. Furthermore, I think it is more readable now.