Extend TAP::Archive recognition to octet-stream
Sometimes the mimetype is recognized as
application/octet-stream, not x-compressed-tar or x-gzip
what it actually really is. We relax this a bit by also
handle octet-stream as TAP::Archive.

It's not that critical because it mostly needs to
know the fact that is *is* an archive in contrast
to text TAP. Later we can extend the actual filetype
recognition if there really should ever be another
filetype for TAP::Archives.