[TEST] for the level2 receiver to Codespeed
The test uses a Tapper config with Codespeed URL
on localhost and a prepared TAP containing data chunks
for codespeed.

It sets up a Tapper::Reports::Receiver and a fake
codespeed instance (just a HTTP::Daemon).

It then sends to the reports receiver and checks whether
subsequent requests appear at the codespeed fake instance.