[FIX] handle exception in processing report
When an exception is thrown while we process the received report this
will skip the exit in the child process and jump back into AnyEvent. Now
we have two AnyEvent processes running.
Since the exception might be generated by Log4Perl we need to report the
exception with a safe log facility. I chose Syslog.